Renamed login controller and put inside controllers folder to conform to MVC structure
